mirror all model route updates to Dispatcher
The current implementation marks Dispatcher as NodeType::DISPATCHER_NODE. But
inside Scheduler, it just treats Dispatcher as a special type of Frontend. In
other words, inside Scheduler, Dispatcher shares as much code with Frontend as
possible. When Scheduler sends ModelRouteProto update to Frontend (e.g. on
LoadModel() and UpdateModelRoutes() RPC), it also sends them to Dispatcher.