ario bench: pass max_flying as a parameter; bugfix
For SEND/RECV, the previous behavior of the bencher is to send out
another message once a previous message is sent. The problem is that
when the reply message is much larger than the request message, the
server's Send Queue will be exhausted. The fix is to let the bencher to
send out another message after a reply is received.