rankmt: simplify RankThread logic
The invariant used to be that
    candidate_pool.size() == num_idle_backends
Hence we needed lots of code to mantain this invariant.
Now, candidate_pool.size() is always the same as the number of model
sessions. Backends are assigned when an active plan triggers timer.
When there is no idle backend available, simple reject the active plan
and wait for ModelThread to provide a new candidate.